netdev-dpdk: Start the device only once on port-add.
When a netdev-dpdk is added to the userspace datapath, rte_eth_start()
is called twice: once from netdev_open() and immediately after from
netdev_reconfigure().

This is unnecessary and wasteful.  A better solution is to postpone
dpdk_eth_dev_init() to the first reconfiguration.  This reduces some
code duplication and makes adding a port faster (~900ms before the
patch, ~400ms after).

Another reason why this is useful is that some DPDK driver might have
problems with reconfiguration. For example, until DPDK commit
8618d19b52b1("net/vmxnet3: reallocate shared memzone on re-config"),
vmxnet3 didn't support being restarted with a different number of
queues.

Technically, the netdev interface changed because before opening rxqs or
calling netdev_send() the user must check if reconfiguration is
required.  This patch also documents that, even though no change to the
userspace datapath (the only user) is required.

Lastly, this patch makes sure the errors returned by ofproto_port_add
(which includes the first port reconfiguration) are reported back to the
database.

Signed-off-by: Daniele Di Proietto <diproiettod@vmware.com>